Refactor NMP
