+++ sporkfish/searcher/negamax.py
@@ -61,7 +61,7 @@ class NegamaxSp(MiniMaxVariants):
         # Null move pruning - reduce the search space by trying a null move,
         # then seeing if the score of the subtree search is still high enough to cause a beta cutoff
         if self._searcher_config.enable_null_move_pruning and self._null_move_pruning(
-            board, depth, alpha, beta
+            board, depth, alpha, beta, self._negamax
         ):
             self._statistics.increment_pruning()
             return beta
@@ -90,9 +90,8 @@ class NegamaxSp(MiniMaxVariants):
             board.push(move)
 
             # Futility pruning
-            if (
-                self._searcher_config.enable_futility_pruning
-                and self._futility_pruning(board, depth, capture, move, alpha)
+            if self._searcher_config.enable_futility_pruning and self._futility_pruning(
+                board, depth, capture, move, alpha
             ):
                 board.pop()
                 self._statistics.increment_pruning()
@@ -128,40 +127,6 @@ class NegamaxSp(MiniMaxVariants):
             self._transposition_table.store(zobrist_state.zobrist_hash, depth, value)
 
         return value
-
-    def _null_move_pruning(
-        self, board: Board, depth: int, alpha: float, beta: float
-    ) -> bool:
-        """
-        Implements null move pruning, a technique to reduce the search space by attempting a 'null move'.
-        It evaluates whether skipping a move (null move) would still allow achieving a beta cutoff,
-        thereby avoiding unnecessary exploration of certain branches of the game tree.
-
-        :param board: The current state of the chess board.
-        :type board: chess.Board
-        :param depth: The current depth in the search tree.
-        :type depth: int
-        :param alpha: The current best score for the maximizing player.
-        :type alpha: float
-        :param beta: The current best score for the minimizing player.
-        :type beta: float
-
-        :return: True if the null move leads to a beta cutoff, indicating a possible pruning opportunity.
-        :rtype: bool
-        """
-        # TODO: add zugzwang check
-        # Will make depth_reduction_factor configurable later
-        depth_reduction_factor = 3
-        in_check = board.is_check()
-        if depth >= depth_reduction_factor and not in_check:
-            null_move_depth = depth - depth_reduction_factor
-            board.push(chess.Move.null())
-            # TODO: check if too expensive to calculate Zobrist state here
-            value = -self._negamax(board, null_move_depth, -beta, -alpha, None)
-            board.pop()
-            if value >= beta:
-                return True
-        return False
 
     def _start_search_from_root(
         self,
